+++ plugins/source/azure/client/testing.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/gorilla/mux"
+	"github.com/rs/zerolog"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
@@ -18,11 +19,9 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/billing/armbilling"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/resources/armresources"
-	"github.com/cloudquery/plugin-pb-go/specs"
-	"github.com/cloudquery/plugin-sdk/v3/faker"
-	"github.com/cloudquery/plugin-sdk/v3/plugins/source"
-	"github.com/cloudquery/plugin-sdk/v3/schema"
-	"github.com/rs/zerolog"
+	"github.com/cloudquery/plugin-sdk/v4/faker"
+	"github.com/cloudquery/plugin-sdk/v4/scheduler"
+	"github.com/cloudquery/plugin-sdk/v4/schema"
 )
 
 const TestSubscription = "12345678-1234-1234-1234-123456789000"
@@ -68,7 +67,7 @@ func (c *MockHttpClient) Do(req *http.Request) (*http.Response, error) {
 }
 
 func MockTestHelper(t *testing.T, table *schema.Table, createServices func(*mux.Router) error) {
-	version := "vDev"
+	// version := "vDev"
 	t.Helper()
 	debug = true
 	table.IgnoreInTests = false
@@ -102,57 +101,55 @@ func MockTestHelper(t *testing.T, table *schema.Table, createServices func(*mux.
 	billingPeriod.ID = to.Ptr("/subscriptions/" + TestSubscription + "/providers/Microsoft.Billing/billingPeriods/202205-1")
 
 	l := zerolog.New(zerolog.NewTestWriter(t)).Output(zerolog.ConsoleWriter{Out: os.Stderr, TimeFormat: time.StampMicro}).Level(zerolog.DebugLevel).With().Timestamp().Logger()
-	newTestExecutionClient := func(ctx context.Context, logger zerolog.Logger, spec specs.Source, _ source.Options) (schema.ClientMeta, error) {
-		err := createServices(router)
-		if err != nil {
-			return nil, err
-		}
-		registeredNamespaces := make(map[string]map[string]bool)
-		registeredNamespaces[TestSubscription] = make(map[string]bool)
-		for _, namespace := range namespaces {
-			registeredNamespaces[TestSubscription][namespace] = true
-		}
 
-		resourceGroup := &armresources.ResourceGroup{}
-		err = faker.FakeObject(resourceGroup)
-		if err != nil {
-			return nil, err
-		}
-		resourceGroup.Name = &testResourceGroup
-
-		c := &Client{
-			logger: l,
-			Options: &arm.ClientOptions{
-				ClientOptions: policy.ClientOptions{
-					Transport: mockClient,
-				},
-			},
-			registeredNamespaces: registeredNamespaces,
-			Creds:                creds,
-			subscriptions:        []string{TestSubscription},
-			ResourceGroups: map[string][]*armresources.ResourceGroup{
-				TestSubscription: {resourceGroup},
-			},
-			BillingAccounts: []*armbilling.Account{&legacyAccount, &modernAccount},
-			BillingPeriods: map[string][]*armbilling.Period{
-				TestSubscription: {&billingPeriod},
-			},
-			storageAccountKeys: &sync.Map{},
-			pluginSpec: &Spec{
-				NormalizeIDs: true,
-			},
-		}
-
-		return c, nil
+	err := createServices(router)
+	if err != nil {
+		t.Fatal(err)
+	}
+	registeredNamespaces := make(map[string]map[string]bool)
+	registeredNamespaces[TestSubscription] = make(map[string]bool)
+	for _, namespace := range namespaces {
+		registeredNamespaces[TestSubscription][namespace] = true
 	}
 
-	p := source.NewPlugin(table.Name, version, []*schema.Table{table}, newTestExecutionClient)
-	p.SetLogger(l)
-	source.TestPluginSync(t, p, specs.Source{
-		Name:         "dev",
-		Path:         "cloudquery/dev",
-		Version:      version,
-		Tables:       []string{table.Name},
-		Destinations: []string{"mock-destination"},
-	})
+	resourceGroup := &armresources.ResourceGroup{}
+	err = faker.FakeObject(resourceGroup)
+	if err != nil {
+		t.Fatal(err)
+	}
+	resourceGroup.Name = &testResourceGroup
+
+	c := &Client{
+		logger: l,
+		Options: &arm.ClientOptions{
+			ClientOptions: policy.ClientOptions{
+				Transport: mockClient,
+			},
+		},
+		registeredNamespaces: registeredNamespaces,
+		Creds:                creds,
+		subscriptions:        []string{TestSubscription},
+		ResourceGroups: map[string][]*armresources.ResourceGroup{
+			TestSubscription: {resourceGroup},
+		},
+		BillingAccounts: []*armbilling.Account{&legacyAccount, &modernAccount},
+		BillingPeriods: map[string][]*armbilling.Period{
+			TestSubscription: {&billingPeriod},
+		},
+		storageAccountKeys: &sync.Map{},
+		pluginSpec: &Spec{
+			NormalizeIDs: true,
+		},
+	}
+	sched := scheduler.NewScheduler(scheduler.WithLogger(l))
+	messages, err := sched.SyncAll(context.Background(), c, schema.Tables{table})
+	if err != nil {
+		t.Fatalf("failed to sync: %v", err)
+	}
+
+	records := messages.GetInserts().GetRecordsForTable(table)
+	emptyColumns := schema.FindEmptyColumns(table, records)
+	if len(emptyColumns) > 0 {
+		t.Fatalf("empty columns: %v", emptyColumns)
+	}
 }