+++ plugins/source/facebookmarketing/resources/plugin/plugin.go
@@ -16,6 +16,7 @@ import (
 	"github.com/cloudquery/plugin-sdk/v4/schema"
 	"github.com/cloudquery/plugin-sdk/v4/transformers"
 	"github.com/rs/zerolog"
+	"golang.org/x/exp/maps"
 )
 
 var (
@@ -39,10 +40,7 @@ var customExceptions = map[string]string{
 
 func titleTransformer(table *schema.Table) {
 	if table.Title == "" {
-		exceptions := make(map[string]string)
-		for k, v := range docs.DefaultTitleExceptions {
-			exceptions[k] = v
-		}
+		exceptions := maps.Clone(docs.DefaultTitleExceptions)
 		for k, v := range customExceptions {
 			exceptions[k] = v
 		}
@@ -57,14 +55,16 @@ func titleTransformer(table *schema.Table) {
 
 type Client struct {
 	plugin.UnimplementedDestination
-	schduler   *scheduler.Scheduler
+	scheduler  *scheduler.Scheduler
 	syncClient *client.Client
 	options    plugin.NewClientOptions
+	allTables  schema.Tables
 }
 
 func newClient(ctx context.Context, logger zerolog.Logger, specBytes []byte, options plugin.NewClientOptions) (plugin.Client, error) {
 	c := &Client{
-		options: options,
+		options:   options,
+		allTables: getTables(),
 	}
 	if options.NoConnection {
 		return c, nil
@@ -79,17 +79,16 @@ func newClient(ctx context.Context, logger zerolog.Logger, specBytes []byte, opt
 		return nil, err
 	}
 	c.syncClient = syncClient.(*client.Client)
-	c.schduler = scheduler.NewScheduler(scheduler.WithLogger(logger), scheduler.WithConcurrency(spec.Concurrency))
+	c.scheduler = scheduler.NewScheduler(scheduler.WithLogger(logger), scheduler.WithConcurrency(spec.Concurrency))
 	return c, nil
 }
 
-func (*Client) Close(ctx context.Context) error {
+func (*Client) Close(_ context.Context) error {
 	return nil
 }
 
-func (*Client) Tables(ctx context.Context, options plugin.TableOptions) (schema.Tables, error) {
-	tables := getTables()
-	tables, err := tables.FilterDfs(options.Tables, options.SkipTables, options.SkipDependentTables)
+func (c *Client) Tables(_ context.Context, options plugin.TableOptions) (schema.Tables, error) {
+	tables, err := c.allTables.FilterDfs(options.Tables, options.SkipTables, options.SkipDependentTables)
 	if err != nil {
 		return nil, err
 	}
@@ -100,12 +99,11 @@ func (c *Client) Sync(ctx context.Context, options plugin.SyncOptions, res chan<
 	if c.options.NoConnection {
 		return fmt.Errorf("no connection")
 	}
-	tables := getTables()
-	tables, err := tables.FilterDfs(options.Tables, options.SkipTables, options.SkipDependentTables)
+	tables, err := c.allTables.FilterDfs(options.Tables, options.SkipTables, options.SkipDependentTables)
 	if err != nil {
 		return err
 	}
-	return c.schduler.Sync(ctx, c.syncClient, tables, res)
+	return c.scheduler.Sync(ctx, c.syncClient, tables, res)
 }
 
 func Plugin() *plugin.Plugin {